XSSHunterPro/requirements.txt
# Core Dependencies
aiohttp==3.9.1
beautifulsoup4==4.12.2
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pymongo==4.6.0
pyyaml==6.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0

# Reporting
jinja2==3.1.2
pdfkit==1.0.0
matplotlib==3.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0 